@@ -24,9 +24,9 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   final cropKey = GlobalKey<CropState>();
-  File _file;
-  File _sample;
-  File _lastCropped;
+   File? _file;
+   File? _sample;
+   File? _lastCropped;
 
   @override
   void dispose() {
@@ -58,7 +58,7 @@ class _MyAppState extends State<MyApp> {
     return Column(
       children: <Widget>[
         Expanded(
-          child: Crop.file(_sample, key: cropKey),
+          child: Crop.file(_sample!, key: cropKey),
         ),
         Container(
           padding: const EdgeInsets.only(top: 20.0),
@@ -72,7 +72,7 @@ class _MyAppState extends State<MyApp> {
                   style: Theme.of(context)
                       .textTheme
                       .button
-                      .copyWith(color: Colors.white),
+                      ?.copyWith(color: Colors.white),
                 ),
                 onPressed: () => _cropImage(),
               ),
@@ -88,7 +88,7 @@ class _MyAppState extends State<MyApp> {
     return TextButton(
       child: Text(
         'Open Image',
-        style: Theme.of(context).textTheme.button.copyWith(color: Colors.white),
+        style: Theme.of(context).textTheme.button?.copyWith(color: Colors.white),
       ),
       onPressed: () => _openImage(),
     );
@@ -96,10 +96,10 @@ class _MyAppState extends State<MyApp> {
 
   Future<void> _openImage() async {
     final pickedFile = await ImagePicker().getImage(source: ImageSource.gallery);
-    final file = File(pickedFile.path);
+    final file = File(pickedFile!.path);
     final sample = await ImageCrop.sampleImage(
       file: file,
-      preferredSize: context.size.longestSide.ceil(),
+      preferredSize: context.size?.longestSide.ceil(),
     );
 
     _sample?.delete();
@@ -112,8 +112,8 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> _cropImage() async {
-    final scale = cropKey.currentState.scale;
-    final area = cropKey.currentState.area;
+    final scale = cropKey.currentState?.scale;
+    final area = cropKey.currentState?.area;
     if (area == null) {
       // cannot crop, widget is not setup
       return;
@@ -122,8 +122,8 @@ class _MyAppState extends State<MyApp> {
     // scale up to use maximum possible number of pixels
     // this will sample image in higher resolution to make cropped image larger
     final sample = await ImageCrop.sampleImage(
-      file: _file,
-      preferredSize: (2000 / scale).round(),
+      file: _file!,
+      preferredSize: (2000 / scale!).round(),
     );
 
     final file = await ImageCrop.cropImage(
