@@ -1,13 +1,10 @@
 package com.lykhonis.imagecropexample;
 
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.embedding.android.FlutterActivity;
 
 public class MainActivity extends FlutterActivity {
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
+  // You do not need to override onCreate() in order to invoke
+  // GeneratedPluginRegistrant. Flutter now does that on your behalf.
+
+  // ...retain whatever custom code you had from before (if any).
 }
